refactor(game.service): ♻️ Improve efficiency and readability of methods
A series of improvements have been made to the code of the methods:

- calculateResult()
- setGame()
- updatedScore()

Unnecessary conditionals have been removed, functions have been reorganized, and an object has been used to determine the score instead of switch or if-else statements. Additionally, Angular programming conventions have been applied to order the elements of the service.